refactor(checker): split resolved def env mirror helper
