+++ crates/tsz-checker/src/context/mod.rs
@@ -37,6 +37,7 @@ pub mod lifetime_shells;
 pub use lifetime_shells::{FileSession, LspPersistentCache, SpeculationScope, WorkerContext};
 mod def_mapping;
 mod def_mapping_formatters;
+mod def_mapping_resolved_env;
 mod deferred_flow_env_write;
 mod unresolved_import;
 pub use deferred_flow_env_write::DeferredFlowEnvWrite;
